requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
python-decouple
google-generativeai
django-cors-headers
